refactor: extract cli helper modules and types
- Split cli.ts by extracting flags, help, and shared types modules
- Add TDD evidence requirement to standards and PR template workflow
- Add focused tests for extracted cli helper modules

Made-with: Cursor
