+++ .github/pull_request_template.md
@@ -9,6 +9,15 @@
 - [ ] I used a TDD flow (red/green/refactor) for behavior changes
 - [ ] I kept the PR scope to one logical concern
 
+## TDD Evidence
+
+- [ ] Behavior change: I added/updated a failing test first (Red), then implemented (Green), then refactored.
+- [ ] Refactor/docs/chore only: no behavior change, so Red step is not applicable.
+- Red evidence (test name/command or short note):
+  - <!-- example: `bun test test/cli.test.ts -t "my failing case"` -->
+- Green evidence (what now passes):
+  - <!-- example: `bun run check` -->
+
 ## Test Plan
 
 - [ ] `bun run typecheck`
